refactor: extract code to use view model structure
